production 일 경우에만 이모지를 삭제하고 보내도록 변경

+++ index.js
@@ -88,8 +88,10 @@ client.on('messageCreate', async (msg) => {
         }
       }
 
-      await client.deleteMessage(msg.channel.id, msg.id);
-      await client.createMessage(msg.channel.id, { embed })
+      if (process.env.NODE_ENV === 'production') {
+        await client.deleteMessage(msg.channel.id, msg.id);
+        await client.createMessage(msg.channel.id, { embed })
+      }
     }
   } catch (e) {
     console.error(e);
